Added code to make use of a HPC cluster to split Ripple data files.
Added optional input arguments HPCCmd and HPCInputFilename to make use of a HPC cluster to split Ripple data files in parallel.

Moved code to split data files into a separate function rsCreateObject so that function can be compiled into an executable (the Matlab Compiler does work well with class constructors).

To reduce the complexity of passing arguments to the compiled application, a MAT file is created with the appropriate arguments in MATLAB before the HPC jobs are submitted. The compiled application then loads the MAT file to figure out which arguments were passed to rplsplit.