Removed need to specify marker format as part of the input arguments.
Now tries to figure out the marker format from the data in each object instead of requiring the user to know which marker format was used in the data.

Also rewrote the code at the end to avoid strange errors. Still not entirely clear why this code is needed though.