Removed need to specify marker format as part of the input arguments.
Now tries to figure out the marker format from the data in each object instead of requiring the user to know which marker format was used in the data.

Added code to create a subplot with all the channels on a particular array on a trial-by-trial format. This can also be used to plot all the LFPs per array, together with the joystick analog channels.

Also rewrote the code at the end to avoid strange errors. Still not entirely clear why this code is needed though.