Added code to use the HPC for spike sorting
Added code to use the HPC for spike sorting (specified using the 'HPC' argument).

Added option to sort every channel except those that are marked 0 for electrodes that are not working (specified using the 'SkipMarker' argument).

Changed the spreadsheet filename to CellActivityLog.xlsx.

Now returns a -1 value if an error (for instance the day being processed is not in the spreadsheet) is encountered.