Modified to make the code for executing arbitrary commands a little more general.
Modified to make the code for executing arbitrary commands a little more general. It now just sends 1 string to be evaluated, instead of sending 2, which was specific to the spike sorting use case. By altering the command string slightly, the spike sorting use case can be supported with the new syntax as well.