Removed the analogTime field.
Removed the analogTime field as it needs to be a double precision variable otherwise the later timepoints become bunched together because of insufficient precision. Since this field can be regenerated when necessary, we can save disk space and reduce load times for the rplhighpass objects if we remove this field.