Added code to correct Unity timestamps using Ripple timestamps.
Added code to correct Unity timestamps using Ripple timestamps.

Created a new field sTime (session time) that starts with Ripple timestamps, and mixes in Unity frame intervals, to create a timeline that we can use to look for spikes.

If the discrepancy in the trial duration in Unity and Ripple are small, we will try to adjust the Unity frame intervals to fit the Ripple duration. If the discrepancy is too large, we will skip that trial in the histogram computation.

Now also computes o_i used in the SIC and MI calculations so that it does not have to re-computed by each cell.

Fixed bug in setIndex that keeps track of which data goes with with session.